Optimize SSR by sending depth pyramid mip offsets via a structured buffer rather than cbuffer
This goes around a quirk in the consoles' HLSL compiler that caused over allocation of VGPRs.

With this change on PS4:
VGPRs:  44 -> 28
SGPRs: 54+2 -> 46+2
Timing (64spp): 7.819ms -> 6.193ms

On Xbox One X
VGPRs: 36 -> 32
SGPRs: 64+2 -> 48+2
Timing (64spp): 14.683 -> 7.835

This increases the fetch cost, but we are never bound by it on this pass.